static_cast of size_t  to int, and Cartesian -> Simple_cartesian

+++ Approximate_min_ellipsoid_d/include/CGAL/Approximate_min_ellipsoid_d.h
@@ -25,7 +25,7 @@
 #include <vector>
 #include <iostream>
 
-#include <CGAL/Cartesian.h>
+#include <CGAL/Simple_cartesian.h>
 #include <CGAL/Approximate_min_ellipsoid_d/Khachiyan_approximation.h>
 
 namespace CGAL {
@@ -137,7 +137,7 @@ namespace CGAL {
       
       // Find e (1+e_eps)-approximation for the embedded points.  This
       // only works when the points affinely span R^{d+1}.
-      E = new Khachiyan_approximation<true,Traits>(d,P.size(),tco);
+      E = new Khachiyan_approximation<true,Traits>(d, static_cast<int>(P.size()),tco);
       const bool is_deg = !E->add(P.begin(),P.end(),e_eps);
 
       // debugging:
